+++ src/pages/meetings/meetingsAddEdit/MeetingsAddEdit.tsx
@@ -1,17 +1,18 @@
-import { useEffect } from 'react';
-import { useState } from 'react';
-import { ReactElement } from 'react';
+import moment from 'moment';
+import { ReactElement, useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { Loader } from '../../../components/loader/Loader';
-import { addMeeting } from '../../../services/meetings.service';
+import { addMeeting, getMeeting, updateMeeting } from '../../../services/meetings.service';
 import { getCurrencies } from '../../../services/selects.service';
 import { getStudentsNames } from '../../../services/students.service';
 import { getTutorsNames } from '../../../services/tutors.service';
+import { MeetingsAddEditView } from './MeetingsAddEditView';
 
-import { MeetingsAddView } from './MeetingsAddView';
+export const MeetingsAddEdit = ({ match }): ReactElement => {
+  const { id } = match.params;
+  const isEdit: boolean = !!id;
 
-export const MeetingsAdd = (): ReactElement => {
   const history = useHistory();
 
   // State
@@ -19,15 +20,30 @@ export const MeetingsAdd = (): ReactElement => {
   const [studentNamesState, setStudentNamesState] = useState([]);
   const [tutorNamesState, setTutorNamesState] = useState([]);
   const [currenciesState, setCurrenciesState] = useState([]);
+  const [meetingState, setMeetingState] = useState({
+    tutorId: '',
+    studentId: '',
+    periodStart: moment(new Date()),
+    periodEnd: moment(new Date()).add(50, 'minutes'),
+    currency: 'GBP',
+    didShow: false,
+    didFillTutorEvaluationSheet: false,
+    didFillStudentEvaluationSheet: false,
+    topicsCovered: '',
+    notes: '',
+  });
 
   useEffect(() => {
     (async () => {
       try {
-        const [tutorNames, studentsNames, currencies] = await Promise.all([
+        const promisesArr = [
           getTutorsNames(),
           getStudentsNames(),
           getCurrencies(),
-        ]);
+        ];
+        if(isEdit) promisesArr.push(getMeeting(id));
+        const [tutorNames, studentsNames, currencies, meeting] =
+          await Promise.all(promisesArr);
         if (tutorNames.success) {
           setTutorNamesState(tutorNames.payload.docs);
         }
@@ -37,13 +53,25 @@ export const MeetingsAdd = (): ReactElement => {
         if (currencies.success) {
           setCurrenciesState(currencies.payload);
         }
+        if (isEdit && meeting.success) {
+          // Update period start and end to normal date format
+          const updatedMeeting = {...meeting.payload};
+          updatedMeeting.periodStart = moment.unix(updatedMeeting.periodStart);
+          updatedMeeting.periodEnd = moment.unix(updatedMeeting.periodEnd);
+          setMeetingState(updatedMeeting);
+        }
         setIsLoadingState(false);
       } catch (error) {
-        console.error('ERROR - MeetingsAdd.tsx - useEffect():', error);
+        console.error('ERROR - MeetingsAddEdit.tsx - useEffect():', error);
         setIsLoadingState(false);
       }
     })();
   }, []);
+
+  const addEditMeeting = async (id: string, values: any) => {
+    if (isEdit) return await updateMeeting(id, values);
+    else return await addMeeting(values);
+  }
 
   const onSubmit = async (values) => {
     try {
@@ -54,14 +82,16 @@ export const MeetingsAdd = (): ReactElement => {
       newValues.periodStart = newValues.periodStart.unix();
       newValues.periodEnd = newValues.periodEnd.unix();
 
-      const newMeeting = await addMeeting(newValues);
+      console.log(newValues);
+      
+      const newMeeting = await addEditMeeting(id, newValues);
       setIsLoadingState(false);
       if (newMeeting.success) {
         history.push('/meetings');
-        toast.success('Added Meeting');
+        toast.success(isEdit ? 'Updated Meeting' : 'Added Meeting');
       }
     } catch (error) {
-      console.error('ERROR - MeetingsAdd.tsx - onSubmit():', error);
+      console.error('ERROR - MeetingsAddEdit.tsx - onSubmit():', error);
       setIsLoadingState(false);
     }
   };
@@ -70,11 +100,12 @@ export const MeetingsAdd = (): ReactElement => {
     <Loader
       isLoading={isLoadingState}
       component={
-        <MeetingsAddView
-          onSubmit={onSubmit}
+        <MeetingsAddEditView
+          meeting={meetingState}
           tutorNames={tutorNamesState}
           studentNames={studentNamesState}
           currencies={currenciesState}
+          onSubmit={onSubmit}
         />
       }
     />
