+++ src/routing/Routes.tsx
@@ -15,13 +15,12 @@ import { Logout } from '../pages/logout/Logout';
 import { ClassesEdit } from '../pages/classes/classesEdit/ClassesEdit';
 import { ClassesAdd } from '../pages/classes/classesAdd/ClassesAdd';
 import { ClassesRead } from '../pages/classes/classesRead/ClassesRead';
-import { MeetingsAdd } from '../pages/meetings/meetingsAdd/MeetingsAdd';
-import { MeetingsEdit } from '../pages/meetings/meetingsEdit/MeetingsEdit';
 import { MeetingsRead } from '../pages/meetings/meetingsRead/MeetingsRead';
 import { StudentsRead } from '../pages/students/studentsRead/StudentsRead';
 import { TutorsRead } from '../pages/tutors/tutorsRead/TutorsRead';
 import { TutorsAddEdit } from '../pages/tutors/tutorsAddEdit/TutorsAddEdit';
 import { StudentsAddEdit } from '../pages/students/studentsAddEdit/StudentsAddEdit';
+import { MeetingsAddEdit } from '../pages/meetings/meetingsAddEdit/MeetingsAddEdit';
 
 export const Routes = () => {
   return (
@@ -36,9 +35,9 @@ export const Routes = () => {
 
       {/* Meetings */}
       <AuthenticatedRoute exact path="/meetings" component={Meetings} />
-      <AuthenticatedRoute exact path="/meetings/edit/:id" component={MeetingsEdit} />
       <AuthenticatedRoute exact path="/meetings/read/:id" component={MeetingsRead} />
-      <AuthenticatedRoute exact path="/meetings/add" component={MeetingsAdd} />
+      <AuthenticatedRoute exact path="/meetings/add" component={MeetingsAddEdit} />
+      <AuthenticatedRoute exact path="/meetings/edit/:id" component={MeetingsAddEdit} />
 
       {/* Students */}
       <AuthenticatedRoute exact path="/students" component={Students} />
